Moved InputMap to its own file
